Flytt alle restrequester som er avhengig av fagsak til FagsakController.kt. Flytt alle restrequester som er avhengig av vedtak til VedtakController.kt
