+++ src/main/kotlin/no/nav/familie/ba/sak/behandling/vedtak/VedtakController.kt
@@ -10,12 +10,19 @@ import no.nav.familie.ba.sak.behandling.steg.BehandlerRolle
 import no.nav.familie.ba.sak.behandling.steg.StegService
 import no.nav.familie.ba.sak.common.RessursUtils.notFound
 import no.nav.familie.ba.sak.sikkerhet.TilgangService
+import no.nav.familie.ba.sak.simulering.SimuleringService
+import no.nav.familie.ba.sak.simulering.domene.RestVedtakSimulering
+import no.nav.familie.ba.sak.simulering.tilbakekreving.RestTilbakekreving
+import no.nav.familie.ba.sak.simulering.tilbakekreving.TilbakekrevingService
+import no.nav.familie.ba.sak.simulering.vedtakSimuleringMottakereTilRestSimulering
+import no.nav.familie.ba.sak.validering.VedtaktilgangConstraint
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.DeleteMapping
+import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -24,88 +31,35 @@ import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-@RequestMapping("/api/fagsaker")
+@RequestMapping("/api/vedtak")
 @ProtectedWithClaims(issuer = "azuread")
 @Validated
 class VedtakController(
-        private val behandlingService: BehandlingService,
         private val vedtakService: VedtakService,
         private val fagsakService: FagsakService,
         private val stegService: StegService,
-        private val tilgangService: TilgangService
+        private val tilbakekrevingService: TilbakekrevingService,
+        private val simuleringService: SimuleringService,
 ) {
-
-    @PostMapping(path = ["/{fagsakId}/vedtak/fritekster"])
-    fun settFritekstVedtakBegrunnelserPåVedtaksperiodeOgType(@PathVariable fagsakId: Long,
-                                                             @RequestBody
-                                                             restPostFritekstVedtakBegrunnelser: RestPostFritekstVedtakBegrunnelser): ResponseEntity<Ressurs<RestFagsak>> {
-        tilgangService.verifiserHarTilgangTilHandling(minimumBehandlerRolle = BehandlerRolle.SAKSBEHANDLER,
-                                                      handling = "legge til vedtakbegrunnelser")
-
-        vedtakService.settFritekstbegrunnelserPåVedtaksperiodeOgType(fagsakId = fagsakId,
-                                                                     restPostFritekstVedtakBegrunnelser = restPostFritekstVedtakBegrunnelser)
-
-        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId))
+    @GetMapping(path = ["/{vedtakId}/simulering"])
+    fun hentSimulering(@PathVariable @VedtaktilgangConstraint
+                       vedtakId: Long): ResponseEntity<Ressurs<RestVedtakSimulering>> {
+        val vedtakSimuleringMottaker = simuleringService.hentEllerOppdaterSimuleringPåVedtak(vedtakId)
+        val restSimulering = vedtakSimuleringMottakereTilRestSimulering(vedtakSimuleringMottaker)
+        return ResponseEntity.ok(Ressurs.success(restSimulering))
     }
 
+    @PostMapping(path = ["/{vedtakId}/tilbakekreving"])
+    fun bekreftSimulering(@PathVariable vedtakId: Long,
+                          @RequestBody restTilbakekreving: RestTilbakekreving?): ResponseEntity<Ressurs<RestFagsak>> {
+        if (restTilbakekreving != null) {
+            tilbakekrevingService.lagreTilbakekreving(restTilbakekreving)
+        }
 
-    @PostMapping(path = ["/{fagsakId}/vedtak/begrunnelser"])
-    fun leggTilVedtakBegrunnelse(@PathVariable fagsakId: Long,
-                                 @RequestBody
-                                 restPostVedtakBegrunnelse: RestPostVedtakBegrunnelse): ResponseEntity<Ressurs<RestFagsak>> {
-        tilgangService.verifiserHarTilgangTilHandling(minimumBehandlerRolle = BehandlerRolle.SAKSBEHANDLER,
-                                                      handling = "legge til vedtakbegrunnelser")
+        val behandling = vedtakService.hent(vedtakId).behandling
+        stegService.håndterSimulering(behandling)
 
-        vedtakService.leggTilVedtakBegrunnelse(fagsakId = fagsakId,
-                                               restPostVedtakBegrunnelse = restPostVedtakBegrunnelse)
-
-        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId))
-    }
-
-    @DeleteMapping(path = ["/{fagsakId}/vedtak/begrunnelser/perioder-vedtaksbegrunnelsetyper"])
-    fun slettVedtakBegrunnelserForPeriodeOgVedtaksbegrunnelseTyper(@PathVariable fagsakId: Long,
-                                                                   @RequestBody
-                                                                   restDeleteVedtakBegrunnelser: RestDeleteVedtakBegrunnelser): ResponseEntity<Ressurs<RestFagsak>> {
-        tilgangService.verifiserHarTilgangTilHandling(minimumBehandlerRolle = BehandlerRolle.SAKSBEHANDLER,
-                                                      handling = "slette vedtakbegrunnelser for periode")
-
-        vedtakService.slettBegrunnelserForPeriodeOgVedtaksbegrunnelseTyper(restDeleteVedtakBegrunnelser = restDeleteVedtakBegrunnelser,
-                                                                           fagsakId = fagsakId)
-
-        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId))
-    }
-
-    @DeleteMapping(path = ["/{fagsakId}/vedtak/begrunnelser/{vedtakBegrunnelseId}"])
-    fun slettVedtakBegrunnelse(@PathVariable fagsakId: Long,
-                               @PathVariable
-                               vedtakBegrunnelseId: Long): ResponseEntity<Ressurs<RestFagsak>> {
-        tilgangService.verifiserHarTilgangTilHandling(minimumBehandlerRolle = BehandlerRolle.SAKSBEHANDLER,
-                                                      handling = "slette vedtakbegrunnelser")
-
-        vedtakService.slettBegrunnelse(fagsakId = fagsakId,
-                                       begrunnelseId = vedtakBegrunnelseId)
-
-        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId))
-    }
-
-    @PostMapping(path = ["/{fagsakId}/send-til-beslutter"], produces = [MediaType.APPLICATION_JSON_VALUE])
-    fun sendBehandlingTilBeslutter(@PathVariable fagsakId: Long,
-                                   @RequestParam behandlendeEnhet: String): ResponseEntity<Ressurs<RestFagsak>> {
-        val behandling = behandlingService.hentAktivForFagsak(fagsakId)
-                         ?: return notFound("Fant ikke behandling på fagsak $fagsakId")
-
-        stegService.håndterSendTilBeslutter(behandling, behandlendeEnhet)
-        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId))
-    }
-
-    @PostMapping(path = ["/{fagsakId}/iverksett-vedtak"], produces = [MediaType.APPLICATION_JSON_VALUE])
-    fun iverksettVedtak(@PathVariable fagsakId: Long,
-                        @RequestBody restBeslutningPåVedtak: RestBeslutningPåVedtak): ResponseEntity<Ressurs<RestFagsak>> {
-        val behandling = behandlingService.hentAktivForFagsak(fagsakId)
-                         ?: return notFound("Fant ikke behandling på fagsak $fagsakId")
-
-        stegService.håndterBeslutningForVedtak(behandling, restBeslutningPåVedtak)
-        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId))
+        return ResponseEntity.ok(fagsakService.hentRestFagsak(fagsakId = behandling.fagsak.id))
     }
 }
